backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
backoff==2.2.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
Django==6.0.1
djangorestframework==3.16.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.4
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langfuse==3.12.1
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
ollama==0.6.1
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
pillow==12.1.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.2
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.5.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
